+++ client/src/book-editor/BookEditor.js
@@ -1,133 +1,15 @@
 import React, { useState, useEffect, useRef } from 'react'
-import PropTypes from 'prop-types'
+import { TransitionGroup, CSSTransition } from 'react-transition-group'
 import { useParams, useHistory } from 'react-router-dom'
+import PropTypes from 'prop-types'
 import io from 'socket.io-client'
-import { LoadingHome } from '../app/App'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import './BookEditor.css'
+import { LoadingHome } from '../app/App'
 import { doCopy, chunkCopytext } from './helper'
 import { EditorNavbar } from './BookEditorNavbar'
-import './BookEditor.css'
 import { CommandChunk } from './CommandChunk'
-
-///////////////////////////////////////////////////////////////////////////////
-// Spark connection handling
-///////////////////////////////////////////////////////////////////////////////
-
-const connectionStatusList = {
-    disconnected: {
-        class: 'disconnected',
-        showControls: true,
-        ready: false,
-        isConnected: false,
-        canDisconnect: false
-    },
-    connecting: {
-        class: 'connecting',
-        showControls: false,
-        ready: false,
-        isConnected: false,
-        canDisconnect: false
-    },
-    connected: {
-        class: 'connected',
-        showControls: false,
-        ready: true,
-        isConnected: true,
-        canDisconnect: true
-    },
-    running: {
-        class: 'running',
-        showControls: false,
-        ready: false,
-        isConnected: true,
-        canDisconnect: true
-    },
-}
-
-// Handles the connection, disconnection and running of chunks
-function useSparkConnection(chunks) {
-    const [connectionStatus, setConnectionStatus] = useState(connectionStatusList.disconnected)
-    const [runningNow, setRunningNow] = useState(null)
-    const [runningId, setRunningId] = useState(null)
-    const [forceRun, setForceRun] = useState(null)
-    const [commandsToRun, setCommandsToRun] = useState([])
-    const sparkSocketRef = useRef(null)
-
-    // Disconnects when leaving the component
-    useEffect(() => {
-        return () => {
-            if (sparkSocketRef.current) {
-                sparkSocketRef.current.disconnect()
-            }
-        }
-    }, [])
-
-    useEffect(() => {
-        if (chunks) {
-            const index = chunks.findIndex(chunk => chunk._id === runningId)
-            setRunningNow(index === -1 ? null : index)
-        } else {
-            setRunningNow(null)
-        }
-    }, [chunks, runningId])
-
-    const connect = (executors, cores, memory) => {
-        sparkSocketRef.current = io(`/spark?executors=${executors}&cores=${cores}&memory=${memory}`)
-        setConnectionStatus(connectionStatusList.connecting)
-
-        sparkSocketRef.current.on('ready', () => {
-            setConnectionStatus(connectionStatusList.connected)
-        })
-    }
-
-    const disconnect = () => {
-        if (sparkSocketRef.current) {
-            sparkSocketRef.current.disconnect()
-        }
-        setRunningNow(null)
-        setCommandsToRun([])
-        setForceRun(null)
-        sparkSocketRef.current = null
-        setConnectionStatus(connectionStatusList.disconnected)
-    }
-
-    const run = (index, command) => {
-        if (sparkSocketRef.current) {
-            sparkSocketRef.current.once('return', () => {
-                if (commandsToRun.length > 0) {
-                // Run until part
-                    setForceRun(commandsToRun.shift())
-                    setCommandsToRun(commandsToRun)
-                } else {
-                // Normal run
-                    setRunningId(null)
-                    setConnectionStatus(connectionStatusList.connected)
-                }
-            })
-            sparkSocketRef.current.emit('run', command)
-            setRunningId(chunks[index]._id)
-            setConnectionStatus(connectionStatusList.running)
-        }
-    }
-
-    const runAllAbove = index => {
-        const runList = chunks.slice(0, index).map(chunk => chunk._id)
-        setForceRun(runList.shift())
-        setCommandsToRun(runList)
-    }
-
-    return {
-        socket: sparkSocketRef.current,
-        status: connectionStatus,
-        runningNow,
-        commandsToRun,
-        forceRun,
-        connect,
-        disconnect,
-        run,
-        runAllAbove
-    }
-}
+import { useSparkConnection } from './useSparkConnection'
 
 ///////////////////////////////////////////////////////////////////////////////
 // Button to add new chunk at a given position
@@ -175,7 +57,10 @@ export function BookEditor() {
         })
 
         // Leave teh current book as ordered
-        bookSocketRef.current.on('exit', () => history.push('/'))
+        bookSocketRef.current.on('logoff', () => {
+            console.log('Teste')
+            history.push('/')
+        })
 
         return () => {
             console.debug('Disconnecting from book socket')
@@ -195,14 +80,22 @@ export function BookEditor() {
         <SparkContext.Provider value={sparkConnection}>
             <LoadingHome loading={loading}>
                 <EditorNavbar book={book} copyAll={copyAll} socket={bookSocketRef.current} />
-                {((book && book.commands) || []).map((chunk, index) => (
-                    <CommandChunk
-                        chunk={chunk}
-                        key={chunk._id}
-                        index={index}
-                        bookSocket={bookSocketRef.current}
-                    />
-                ))}
+                <TransitionGroup>
+                    {((book && book.commands) || []).map((chunk, index) => (
+                        <CSSTransition
+                            key={chunk._id}
+                            classNames="chunk-animate"
+                            timeout={500}
+                            unmountOnExit
+                        >
+                            <CommandChunk
+                                chunk={chunk}
+                                index={index}
+                                bookSocket={bookSocketRef.current}
+                            />
+                        </CSSTransition>
+                    ))}
+                </TransitionGroup>
                 <ChunkAddButton bookSocket={bookSocketRef.current} />
             </LoadingHome>
         </SparkContext.Provider>