+++ client/src/book-editor/BookEditorNavbar.js
@@ -1,94 +1,13 @@
-import React, { useContext, useRef, useEffect } from 'react'
+import React, { useRef, useEffect, useState } from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
-import { Navbar, Dropdown, Button, Form, InputGroup } from 'react-bootstrap'
+import { Navbar, Dropdown } from 'react-bootstrap'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { SparkContext } from './BookEditor'
 import { editablePrevent } from './helper'
 import { SimpleDropdown } from '../components/simple-dropdown/SimpleDropdown'
+import { SparkConnectionControl } from './SparkConnectionControl'
+import { ConsoleModal } from './ConsoleModal'
 import './BookEditorNavbar.css'
-
-///////////////////////////////////////////////////////////////////////////////
-// Connection widget
-///////////////////////////////////////////////////////////////////////////////
-ConnectionControl.propTypes = {
-    config: PropTypes.exact({
-        executors: PropTypes.number,
-        cores: PropTypes.number,
-        memory: PropTypes.number,
-    }),
-    socket: PropTypes.object
-}
-function ConnectionControl({ config, socket }) {
-    const spark = useContext(SparkContext)
-    const executors = useRef()
-    const cores = useRef()
-    const memory = useRef()
-
-    useEffect(() => {
-        if (config) {
-            if (executors.current) {
-                executors.current.value = config.executors
-            }
-            if (cores.current) {
-                cores.current.value = config.cores
-            }
-            if (memory.current) {
-                memory.current.value = config.memory
-            }
-        }
-    }, [config])
-
-    const connect = () => {
-        spark.connect(executors.current.value, cores.current.value, memory.current.value)
-    }
-
-    const changeConfig = () => {
-        socket.emit('spark.config', executors.current.value, cores.current.value, memory.current.value)
-    }
-
-    const controls = [
-        { name: 'Executors', ref: executors },
-        { name: 'Cores', ref: cores },
-        { name: 'Memory', ref: memory }
-    ]
-
-    return (
-        <Form as="div" inline className={`p-2 connection-control ${spark.status.class}`}>
-            { spark.status.showControls ? controls.map((control, index) => (
-                <InputGroup size="sm" className="mr-2" key={index}>
-                    <InputGroup.Prepend>
-                        <InputGroup.Text>{control.name}</InputGroup.Text>
-                    </InputGroup.Prepend>
-                    <Form.Control ref={control.ref} onChange={changeConfig} type="number" defaultValue="2" className="text-center" min="1" style={{ width: '4em' }}></Form.Control>
-                </InputGroup>
-            )).concat((
-                <Button size="sm" className="mr-2" onClick={connect} key="connectButton">Connect</Button>
-            )) : null }
-            { spark.runningNow != null ? (
-                <div className="mr-2">
-                    <FontAwesomeIcon icon="spinner" className="mr-2 text-primary rotate" />
-                    Running Chunk { spark.runningNow }
-                </div>
-            ) : null }
-            <Dropdown drop="left">
-                <Dropdown.Toggle as={SimpleDropdown}>
-                    <div className="connection-icon">
-                        <FontAwesomeIcon icon="wifi" />
-                    </div>
-                </Dropdown.Toggle>
-                { spark.status.canDisconnect ? (
-                    <Dropdown.Menu>
-                        <Dropdown.Item onClick={spark.disconnect}>
-                            <FontAwesomeIcon icon="power-off" className="mr-2" />
-                            Disconnect
-                        </Dropdown.Item>
-                    </Dropdown.Menu>
-                ) : null }
-            </Dropdown>
-        </Form>
-    )
-}
 
 ////////////////////////////////////////////////////////////////////////////////
 // Book Name
@@ -142,6 +61,15 @@ EditorNavbar.propTypes = {
     socket: PropTypes.object
 }
 export function EditorNavbar({ book, copyAll, socket }) {
+    const [showConsole, setShowConsole] = useState(false)
+
+    const openConsole = () => {
+        setShowConsole(true)
+    }
+
+    const closeConsole = () => {
+        setShowConsole(false)
+    }
 
     return (
         <Navbar variant="dark" className="sticky-top d-flex shadow">
@@ -153,11 +81,13 @@ export function EditorNavbar({ book, copyAll, socket }) {
             </Navbar.Brand>
             <div className="flex-grow-1"/>
             <div className="mr-2">
-                <ConnectionControl config={book.sparkConfig} socket={socket}/>
+                <SparkConnectionControl config={book.sparkConfig} socket={socket}/>
             </div>
             <Dropdown drop="left">
                 <Dropdown.Toggle as={SimpleDropdown}>
-                    <FontAwesomeIcon icon="ellipsis-v"/>
+                    <div className="pl-2">
+                        <FontAwesomeIcon icon="ellipsis-v"/>
+                    </div>
                 </Dropdown.Toggle>
                 <Dropdown.Menu >
                     <Dropdown.Item onClick={copyAll}>
@@ -168,8 +98,13 @@ export function EditorNavbar({ book, copyAll, socket }) {
                         <FontAwesomeIcon icon="file-download" className="mr-2" />
                         Download
                     </Dropdown.Item>
+                    <Dropdown.Item onClick={openConsole}>
+                        <FontAwesomeIcon icon="terminal" className="mr-2" />
+                        Console
+                    </Dropdown.Item>
                 </Dropdown.Menu>
             </Dropdown>
+            <ConsoleModal show={showConsole} close={closeConsole} />
         </Navbar>
     )
 }