+++ server/router-api.js
@@ -1,8 +1,10 @@
 const express = require('express')
 const router = express.Router()
 
-router.use(require('./api/login-api').accessControl)
-router.use('/login', require('./api/login-api').router)
+const loginApi = require('./api/login-api')
+
+router.use(loginApi.accessControlMiddleware)
+router.use('/login', loginApi.router)
 router.use('/book', require('./api/book-api'))
 
 module.exports = router